+++ packages/core/page-layout/src/composables/index.ts
@@ -1,0 +1,5 @@
+import useI18n from './useI18n'
+
+export default {
+  useI18n,
+}