+++ packages/core/core/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.11.12](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/core@1.11.11...@kong-ui-public/core@1.11.12) (2026-02-12)
+
+**Note:** Version bump only for package @kong-ui-public/core
+
+
+
+
+
 ## [1.11.11](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/core@1.11.10...@kong-ui-public/core@1.11.11) (2026-01-26)