+++ packages/core/documentation/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.51](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/documentation@1.4.50...@kong-ui-public/documentation@1.4.51) (2026-02-12)
+
+**Note:** Version bump only for package @kong-ui-public/documentation
+
+
+
+
+
 ## [1.4.50](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/documentation@1.4.49...@kong-ui-public/documentation@1.4.50) (2026-02-04)
 
 **Note:** Version bump only for package @kong-ui-public/documentation