+++ packages/core/forms/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.21.3](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/forms@4.21.2...@kong-ui-public/forms@4.21.3) (2026-02-12)
+
+**Note:** Version bump only for package @kong-ui-public/forms
+
+
+
+
+
 ## [4.21.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/forms@4.21.1...@kong-ui-public/forms@4.21.2) (2026-02-04)
 
 **Note:** Version bump only for package @kong-ui-public/forms