+++ packages/core/page-layout/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/page-layout@0.1.0...@kong-ui-public/page-layout@0.2.0) (2026-02-25)
+
+
+### Features
+
+* **page-layout:** tabs navbar component [KHCP-19463] ([#2922](https://github.com/Kong/public-ui-components/issues/2922)) ([a7b8d35](https://github.com/Kong/public-ui-components/commit/a7b8d350baee3fbbca387e82809e0b3bb8425ebf))
+
+
+
+
+
 # 0.1.0 (2026-02-05)