+++ packages/entities/entities-keys/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.11.6](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-keys@3.11.5...@kong-ui-public/entities-keys@3.11.6) (2026-02-12)
+
+**Note:** Version bump only for package @kong-ui-public/entities-keys
+
+
+
+
+
 ## [3.11.5](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-keys@3.11.4...@kong-ui-public/entities-keys@3.11.5) (2026-02-04)
 
 **Note:** Version bump only for package @kong-ui-public/entities-keys